Can delete video if hls converting stopped

+++ VideoStreaming/CatalogService/Services/VideoCatalogService.cs
@@ -39,7 +39,10 @@ namespace CatalogService.Services
                     file.Delete();
                 }
                 var hlsDir = new DirectoryInfo(Path.Combine(_fileStorageSettings.Path, "hls", id));
-                hlsDir.Delete(true);
+                if (hlsDir.Exists)
+                {
+                    hlsDir.Delete(true);
+                }
                 _catalogDb.Remove(video);
                 await _catalogDb.SaveChangesAsync();
             }
